#58 Working on GATT Characteristic
