+++ Tests/LinuxMain.swift
@@ -22,6 +22,5 @@ XCTMain([
     testCase(GATTTests.allTests),
     testCase(GATTDatabaseTests.allTests),
     testCase(GATTDescriptorTests.allTests),
-    testCase(GATTAlertNotificationTests.allTests),
-    testCase(GATTBatteryTests.allTests)
+    testCase(GATTCharacteristicTests.allTests)
     ])
